requirements.txt
Django~=3.0.7
psycopg2-binary
imdbpy